*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@46 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/cspfj/MACSolver.java
@@ -81,7 +81,7 @@ public final class MACSolver extends AbstractSolver {
 			return false;
 		}
 
-		final int pair = heuristic.selectPair();
+		final long pair = heuristic.selectPair();
 
 		final Variable selectedVariable = Pair.variable(pair);
