+++ src/cspfj/heuristic/WdJW.java
@@ -140,7 +140,7 @@ public final class WdJW implements Heuristic {
 		recompute.clear();
 	}
 
-	public int selectPair() {
+	public long selectPair() {
 		recompute();
 		Variable bestVariable = null;
 		int bestValue = -1;
